Plugins: Add PinnedPlugins component
