+++ edit-post/store/defaults.js
@@ -5,5 +5,4 @@ export const PREFERENCES_DEFAULTS = {
 	features: {
 		fixedToolbar: false,
 	},
-	pinnedPluginItems: {},
 };
