+++ edit-post/store/test/actions.js
@@ -11,7 +11,6 @@ import {
 	openModal,
 	closeModal,
 	toggleFeature,
-	togglePinnedPluginItem,
 	requestMetaBoxUpdates,
 	initializeMetaBoxState,
 } from '../actions';
@@ -93,17 +92,6 @@ describe( 'actions', () => {
 			expect( toggleFeature( feature ) ).toEqual( {
 				type: 'TOGGLE_FEATURE',
 				feature,
-			} );
-		} );
-	} );
-
-	describe( 'togglePinnedPluginItem', () => {
-		it( 'should return TOGGLE_PINNED_PLUGIN_ITEM action', () => {
-			const pluginName = 'foo/bar';
-
-			expect( togglePinnedPluginItem( pluginName ) ).toEqual( {
-				type: 'TOGGLE_PINNED_PLUGIN_ITEM',
-				pluginName,
 			} );
 		} );
 	} );
