+++ edit-post/store/test/reducer.js
@@ -25,7 +25,6 @@ describe( 'state', () => {
 				isGeneralSidebarDismissed: false,
 				panels: { 'post-status': true },
 				features: { fixedToolbar: false },
-				pinnedPluginItems: {},
 			} );
 		} );
 
@@ -85,42 +84,6 @@ describe( 'state', () => {
 			} );
 
 			expect( state.features ).toEqual( { chicken: false } );
-		} );
-
-		describe( 'pinnedPluginItems', () => {
-			const initialState = deepFreeze( {
-				pinnedPluginItems: {
-					'foo/enabled': true,
-					'foo/disabled': false,
-				},
-			} );
-
-			it( 'should disable a pinned plugin flag when the value does not exist', () => {
-				const state = preferences( initialState, {
-					type: 'TOGGLE_PINNED_PLUGIN_ITEM',
-					pluginName: 'foo/does-not-exist',
-				} );
-
-				expect( state.pinnedPluginItems[ 'foo/does-not-exist' ] ).toBe( false );
-			} );
-
-			it( 'should disable a pinned plugin flag when it is enabled', () => {
-				const state = preferences( initialState, {
-					type: 'TOGGLE_PINNED_PLUGIN_ITEM',
-					pluginName: 'foo/enabled',
-				} );
-
-				expect( state.pinnedPluginItems[ 'foo/enabled' ] ).toBe( false );
-			} );
-
-			it( 'should enable a pinned plugin flag when it is disabled', () => {
-				const state = preferences( initialState, {
-					type: 'TOGGLE_PINNED_PLUGIN_ITEM',
-					pluginName: 'foo/disabled',
-				} );
-
-				expect( state.pinnedPluginItems[ 'foo/disabled' ] ).toBe( true );
-			} );
 		} );
 	} );
