+++ packages/jest-preset-default/jest-preset.json
@@ -21,6 +21,11 @@
 		"**/?(*.)(spec|test).js",
 		"**/test/*.js"
 	],
+	"testPathIgnorePatterns": [
+		"/node_modules/",
+		"<rootDir>/.*/build/",
+		"<rootDir>/.*/build-module/"
+	],
 	"timers": "fake",
 	"transform": {
 		"^.+\\.jsx?$": "<rootDir>/node_modules/babel-jest"
