+++ packages/plugins/package.json
@@ -20,10 +20,15 @@
 	"module": "build-module/index.js",
 	"dependencies": {
 		"@babel/runtime-corejs2": "7.0.0-beta.56",
+		"@wordpress/components": "file:../components",
 		"@wordpress/compose": "file:../compose",
+		"@wordpress/data": "file:../data",
 		"@wordpress/element": "file:../element",
 		"@wordpress/hooks": "file:../hooks",
 		"lodash": "^4.17.10"
+	},
+	"devDependencies": {
+		"deep-freeze": "^0.0.1"
 	},
 	"publishConfig": {
 		"access": "public"
