+++ packages/plugins/src/components/index.js
@@ -1,2 +1,3 @@
+export { default as PinnedPlugins } from './pinned-plugins';
 export { default as PluginArea } from './plugin-area';
 export { withPluginContext } from './plugin-context';
