+++ packages/plugins/src/components/pinned-plugins/style.scss
@@ -1,4 +1,4 @@
-.edit-post-pinned-plugins {
+.plugins-pinned-plugins {
 	display: flex;
 
 	.components-icon-button {
