+++ packages/plugins/src/index.js
@@ -1,2 +1,7 @@
+/**
+ * Internal dependencies
+ */
+import './store';
+
 export * from './components';
 export * from './api';
