+++ packages/plugins/src/store/defaults.js
@@ -1,0 +1,3 @@
+export const PREFERENCES_DEFAULTS = {
+	pinnedPluginItems: {},
+};
