+++ packages/plugins/src/style.scss
@@ -1,0 +1,1 @@
+@import "./components/pinned-plugins/style.scss";
